Final-Lumen/requirements.txt
# FastAPI and Web Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
cryptography==44.0.0

# AI & ML
google-generativeai==0.8.3
scikit-learn==1.6.0
numpy==2.1.3
pandas==2.2.3
faiss-cpu==1.9.0.post1
# chromadb==0.5.23
sentence-transformers==3.3.1

# OCR & Document Processing
pytesseract==0.3.13
pdf2image==1.17.0
Pillow==11.0.0
opencv-python==4.10.0.84

# Text Processing & NLP
# spacy==4.0.0.dev3  # Optional - requires C++ build tools on Windows, skip if installation fails
fuzzywuzzy==0.18.0
python-Levenshtein==0.27.1
rapidfuzz==3.10.1  # Better alternative to fuzzywuzzy, pure Python

# External Integrations
google-api-python-client==2.154.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
twilio==9.3.7
requests==2.32.3
beautifulsoup4==4.12.3

# Utilities
python-dotenv==1.0.1
pytz==2024.2
dateparser==1.2.0
phonenumbers==8.13.50

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
faker==33.1.0

# Logging & Monitoring
python-json-logger==3.2.1
